いろいろ修正
